create fisc str utility
