+++ src/a15_fisc_logic/_utils/str_a15.py
@@ -1,0 +1,76 @@
+def timeline_str() -> str:
+    return "timeline"
+
+
+def offi_time_str() -> str:
+    return "offi_time"
+
+
+def brokerunits_str() -> str:
+    return "brokerunits"
+
+
+def cashbook_str() -> str:
+    return "cashbook"
+
+
+def amount_str() -> str:
+    return "amount"
+
+
+def month_tag_str() -> str:
+    return "month_tag"
+
+
+def hour_tag_str() -> str:
+    return "hour_tag"
+
+
+def cumlative_minute_str() -> str:
+    return "cumlative_minute"
+
+
+def cumlative_day_str() -> str:
+    return "cumlative_day"
+
+
+def weekday_tag_str() -> str:
+    return "weekday_tag"
+
+
+def weekday_order_str() -> str:
+    return "weekday_order"
+
+
+# def fiscunit_str()-> str: return "fiscunit"
+# def fisc_dealunit_str()-> str: return "fisc_dealunit"
+# def fisc_cashbook_str()-> str: return "fisc_cashbook"
+# def fisc_timeline_hour_str()-> str: return "fisc_timeline_hour"
+# def fisc_timeline_month_str()-> str: return "fisc_timeline_month"
+# def fisc_timeline_weekday_str()-> str: return "fisc_timeline_weekday"
+def fiscunit_str() -> str:
+    return "fiscunit"
+
+
+def fisc_dealunit_str() -> str:
+    return "fisc_dealunit"
+
+
+def fisc_cashbook_str() -> str:
+    return "fisc_cashbook"
+
+
+def fisc_timeline_hour_str() -> str:
+    return "fisc_timeline_hour"
+
+
+def fisc_timeline_month_str() -> str:
+    return "fisc_timeline_month"
+
+
+def fisc_timeline_weekday_str() -> str:
+    return "fisc_timeline_weekday"
+
+
+def fisc_timeoffi_str() -> str:
+    return "fisc_timeoffi"
