+++ src/a15_fisc_logic/fisc_config.py
@@ -3,50 +3,6 @@ from src.a00_data_toolbox.dict_toolbox import get_from_nested_dict
 from os import getcwd as os_getcwd
 
 
-def timeline_str() -> str:
-    return "timeline"
-
-
-def offi_time_str() -> str:
-    return "offi_time"
-
-
-def brokerunits_str() -> str:
-    return "brokerunits"
-
-
-def cashbook_str() -> str:
-    return "cashbook"
-
-
-def amount_str() -> str:
-    return "amount"
-
-
-def month_tag_str() -> str:
-    return "month_tag"
-
-
-def hour_tag_str() -> str:
-    return "hour_tag"
-
-
-def cumlative_minute_str() -> str:
-    return "cumlative_minute"
-
-
-def cumlative_day_str() -> str:
-    return "cumlative_day"
-
-
-def weekday_tag_str() -> str:
-    return "weekday_tag"
-
-
-def weekday_order_str() -> str:
-    return "weekday_order"
-
-
 def get_fisc_config_filename() -> str:
     return "fisc_config.json"
 
@@ -54,40 +10,6 @@ def get_fisc_config_filename() -> str:
 def config_file_dir() -> str:
     src_dir = create_path(os_getcwd(), "src")
     return create_path(src_dir, "a15_fisc_logic")
-
-
-# def fiscunit_str()-> str: return "fiscunit"
-# def fisc_dealunit_str()-> str: return "fisc_dealunit"
-# def fisc_cashbook_str()-> str: return "fisc_cashbook"
-# def fisc_timeline_hour_str()-> str: return "fisc_timeline_hour"
-# def fisc_timeline_month_str()-> str: return "fisc_timeline_month"
-# def fisc_timeline_weekday_str()-> str: return "fisc_timeline_weekday"
-def fiscunit_str() -> str:
-    return "fiscunit"
-
-
-def fisc_dealunit_str() -> str:
-    return "fisc_dealunit"
-
-
-def fisc_cashbook_str() -> str:
-    return "fisc_cashbook"
-
-
-def fisc_timeline_hour_str() -> str:
-    return "fisc_timeline_hour"
-
-
-def fisc_timeline_month_str() -> str:
-    return "fisc_timeline_month"
-
-
-def fisc_timeline_weekday_str() -> str:
-    return "fisc_timeline_weekday"
-
-
-def fisc_timeoffi_str() -> str:
-    return "fisc_timeoffi"
 
 
 def get_fisc_config_dict() -> dict:
