+++ src/a17_idea_logic/test__idea_db_tool/test_idea_query_builder.py
@@ -8,7 +8,7 @@ from src.a06_bud_logic._utils.str_a06 import (
     credit_belief_str,
     debtit_belief_str,
 )
-from src.a15_fisc_logic.fisc_config import amount_str
+from src.a15_fisc_logic._utils.str_a15 import amount_str
 from src.a17_idea_logic.idea_config import get_idea_config_dict
 from src.a17_idea_logic.idea_db_tool import (
     create_idea_sorted_table,
