+++ src/a18_etl_toolbox/test_/test_tran_sqlstrs_.py
@@ -35,8 +35,7 @@ from src.a07_calendar_logic._utils.str_a07 import (
     yr1_jan1_offset_str,
 )
 from src.a08_bud_atom_logic.atom_config import get_bud_dimens, get_delete_key_name
-from src.a15_fisc_logic.fisc_config import (
-    get_fisc_dimens,
+from src.a15_fisc_logic._utils.str_a15 import (
     fiscunit_str,
     fisc_cashbook_str,
     fisc_dealunit_str,
@@ -45,6 +44,7 @@ from src.a15_fisc_logic.fisc_config import (
     fisc_timeline_weekday_str,
     fisc_timeoffi_str,
 )
+from src.a15_fisc_logic.fisc_config import get_fisc_dimens
 from src.a16_pidgin_logic.pidgin_config import get_pidgin_dimens
 from src.a17_idea_logic.idea_config import (
     idea_number_str,
