+++ src/a18_etl_toolbox/test_etl_db2obj/test_db2obj_fisc.py
@@ -1,9 +1,10 @@
 from src.a02_finance_logic._utils.strs_a02 import fisc_tag_str
-from src.a15_fisc_logic.fisc import (
-    fiscunit_shop,
-    get_from_dict as fiscunit_get_from_dict,
+from src.a15_fisc_logic.fisc import get_from_dict as fiscunit_get_from_dict
+from src.a15_fisc_logic._utils.str_a15 import (
+    cashbook_str,
+    brokerunits_str,
+    timeline_str,
 )
-from src.a15_fisc_logic.fisc_config import cashbook_str, brokerunits_str, timeline_str
 from src.a18_etl_toolbox.tran_sqlstrs import create_fisc_prime_tables
 from src.a18_etl_toolbox.db_obj_tool import get_fisc_dict_from_db
 from sqlite3 import connect as sqlite3_connect
