+++ src/a19_world_logic/test_00_sound_to_yell/test_world_etl00_sound_to_yell_raw.py
@@ -6,7 +6,7 @@ from src.a00_data_toolbox.db_toolbox import (
 )
 from src.a02_finance_logic._utils.strs_a02 import fisc_tag_str
 from src.a06_bud_logic._utils.str_a06 import face_name_str, event_int_str
-from src.a15_fisc_logic.fisc_config import cumlative_minute_str, hour_tag_str
+from src.a15_fisc_logic._utils.str_a15 import cumlative_minute_str, hour_tag_str
 from src.a17_idea_logic.idea_db_tool import upsert_sheet, yell_raw_str
 from src.a19_world_logic.world import worldunit_shop
 from src.a19_world_logic._utils.env_a19 import (
