+++ src/a19_world_logic/test_etl_pipeline/test_world_sound_to_fisc_unit.py
@@ -1,9 +1,4 @@
-from src.a00_data_toolbox.file_toolbox import (
-    create_path,
-    count_dirs_files,
-    delete_dir,
-    save_file,
-)
+from src.a00_data_toolbox.file_toolbox import create_path, count_dirs_files, save_file
 from src.a02_finance_logic._utils.strs_a02 import (
     owner_name_str,
     fisc_tag_str,
@@ -19,7 +14,7 @@ from src.a12_hub_tools.hub_path import (
     create_deal_acct_mandate_ledger_path as deal_mandate,
     create_fisc_ote1_csv_path,
 )
-from src.a15_fisc_logic.fisc_config import cumlative_minute_str, hour_tag_str
+from src.a15_fisc_logic._utils.str_a15 import cumlative_minute_str, hour_tag_str
 from src.a17_idea_logic.idea_db_tool import (
     upsert_sheet,
     sheet_exists,
